+++ jxls-site/docs/changes.md
@@ -11,8 +11,8 @@ tickets:
 - [#402 more complex direction=RIGHT demo](https://github.com/jxlsteam/jxls/pull/402)
 - [#396 POI bug 66679 compensation](https://github.com/jxlsteam/jxls/issues/396) activate with withCellStyleGeneralEnsurer(),
   also new: template preprocessing
+- [#395 Hyperlink](https://github.com/jxlsteam/jxls/issues/395)
 - [#394 POI 5.5.1](https://github.com/jxlsteam/jxls/pull/394)
-- [#377 Add link command](https://github.com/jxlsteam/jxls/pull/377) *- contribution by [kingwtd](https://github.com/kingwtd)*
 - [#375 Same multisheet sheet name and template sheet name results in removed sheet without any error](https://github.com/jxlsteam/jxls/issues/375)
 - [#371 commons-beanutils 1.11.0](https://github.com/jxlsteam/jxls/issues/371)
 - [#366 PreWriteActions are ignored if KeepTemplateSheet is set to KEEP](https://github.com/jxlsteam/jxls/issues/366)
@@ -30,6 +30,7 @@ tickets:
 - [#195 New area column merge command](https://github.com/jxlsteam/jxls/issues/195)
   *- contribution by [mannixsuo](https://github.com/mannixsuo)*
 - [#63 Set fetch size for DatabaseAccess](https://github.com/jxlsteam/jxls/issues/63)
+- [#46 Data validation (drop down) not displayed in generated fields when using EachCommand (jx:each)](https://github.com/jxlsteam/jxls/issues/46)
 - see test notes in [#242](https://github.com/jxlsteam/jxls/issues/242) and [#394](https://github.com/jxlsteam/jxls/pull/394)
 
 ## v3.0.0 <!-- ** -->